# xapi-types/ref: optimize of_string for real references 