Estimated testing time 20s (2 benchmarks x 10s). Change using '-quota'.
```
┌────────┬──────────┬─────────┬────────────┐
│ Name   │ Time/Run │ mWd/Run │ Percentage │
├────────┼──────────┼─────────┼────────────┤
│ master │  63.58ns │  28.00w │    100.00% │
│ pr     │  37.84ns │  14.00w │     59.51% │
└────────┴──────────┴─────────┴────────────┘
```